refactor(admin): decompose admin.ts into 9 domain modules
Split 4,639-line monolith into focused domain modules:
auth, dashboard, users, matching, messaging, content,
moderation, revenue, system. Barrel export in index.ts
maintains backward compatibility for all existing imports.

Original admin.ts replaced with re-export shim.
