+++ app/services/admin/auth.ts
@@ -1,0 +1,6 @@
+export const auth = {
+	cleanup: () => {
+		localStorage.removeItem('user');
+		localStorage.removeItem('isAdmin');
+	},
+};
